% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmemod.R
\name{rmemod}
\alias{rmemod}
\alias{print.rmemod}
\alias{plot.rmemod}
\title{rmemod}
\usage{
rmemod(
  x,
  modsize = 2L,
  ngenes = nrow(x),
  bgrate = 0.01037848,
  winnow = 4,
  minfreq = 0.1,
  threshold = 50,
  verbose = FALSE,
  outdir = NULL,
  timeout = 30
)

\method{print}{rmemod}(x, ...)

\method{plot}{rmemod}(x, ...)
}
\arguments{
\item{x}{an \code{n x p} \emph{binary} matrix of \code{n} genes and
\code{p} samples or an object returned by \code{rmemod} for print and
plot methods}

\item{modsize}{the largest module that the algorithm will attempt to find,
i.e., potential modules will be between 2 and \code{modsize} genes

warning: the algorithm's complexity grows very quickly as a result of
using combinatorial search, e.g., values over 5 applied to very large
matrices may take a long time}

\item{ngenes}{an integer representing the total number of genes assayed
which allows for multiple testing correction; include all genes assayed
even if they are not represented in \code{x} (there is little reason to
include genes in the matrix that have no mutations in any sample)}

\item{bgrate}{the background mutation rate, i.e., the expected odds of a
particular attribute in a particular sample being altered assuming no
selective pressure; the default value (0.01037848) assumes data composed
of copy number and somatic mutation assays and is derived from HapMap
data and estimation of passenger mutation rates in glioblastoma multiforme}

\item{winnow}{Winnow score threshold; the algorithm speeds up the search
process by excluding poor edges, and this parameter controls the threshold
score for an edge to be kept; due to Winnow's design, these values should
be powers of 2 (4, 8, 16, ...), and if not will be coerced to a power of
2 using \code{2 ^ floor(log2(winnow))} before running algorithm

suggested values:

  \tabular{llll}{
  \tab \code{attributes} \tab \code{samples} \tab \code{winnow} \cr
  \tab \code{1k} \tab \code{200} \tab \code{4} \cr
  \tab \code{5k} \tab \code{500} \tab \code{32} \cr
  \tab \code{18k} \tab \code{500} \tab \code{128} \cr
  }}

\item{minfreq}{minimum frequency of alteration required for a gene to be
included in the search for modules; recommended default is 0.10 as the
false positive rate increases below that point}

\item{threshold}{the minimum threshold for algorithmic significance value
that a module must exceed; optimal values will depend on input data size

suggested values:

  \tabular{llll}{
  \tab \code{genes} \tab \code{samples} \tab \code{threshold} \cr
  \tab \code{1k} \tab \code{200} \tab \code{100} \cr
  \tab \code{5k} \tab \code{500} \tab \code{200} \cr
  \tab \code{18k} \tab \code{500} \tab \code{300} \cr
  }}

\item{verbose}{logical; if \code{TRUE}, stderr/stdout output will be printed}

\item{outdir}{optional directory to save output files (i.e., those created
by \code{rmemod} as data frames); otherwise, temporary files are used
and deleted at the end of the current \code{R} session}

\item{timeout}{timeout in seconds for allowing the algorithm to run; use
0 for no timeout (see \code{\link{system2}}); if the algorithm is still
running after \code{timeout} seconds, the process will be killed and no
results will be returned}

\item{...}{additional arguments passed to or from other methods}
}
\value{
A list of class \code{"rmemode"} with the following elements:

\item{\code{potential}}{a data frame of the complete list of all potential
  modules and the d-score for each}
\item{\code{top}}{a data frame of the largest and most significant non
  overlapping modules which have d-score of at least \code{threshold};
  coverage, exclusivity, and p-value are also included; see details}
\item{\code{call}}{the call used to invoke the algorithm with options}
\item{\code{sys}}{the return value of \code{\link{system2}}, potentially
  warnings or errors, stdout/stderr if \code{verbose = TRUE}, or \code{0}
  otherwise upon successful completion of the algorithm}
}
\description{
Build a weighted graph using the Winnow algorithm and search for modules
that are highly recurrent (across samples) and have high levels of mutual
exclusivity. The significance of these patterns is determined using the
algorithmic significance test.
}
\details{
The d-score is the algorithmic significance value with significance being
equal to \code{2^-d}. Coverage is defined as the percentage of samples that
contain at least one aberration within the module. Exclusivity is defined
as the percentage of covered samples that contain exactly one aberration
within the module.
}
\examples{
## basic usage
set.seed(1)
x <- matrix(rbinom(1000 * 50, 1, 0.1), 100)
r <- rmemod(x)
r
plot(r)


## miller, 2011
f <- system.file(
  'etc', 'rmeMod', 'exampleFiles', 'input.dat',
  package = 'rmemod'
)
x <- as.matrix(read.table(f))
r <- rmemod(x, 3, threshold = 200)
r

## potential modules
head(r$pot)
## significance
head(2 ^ -r$pot$d)

r <- rmemod(x, 3, threshold = 50)
plot(r)

}
\references{
Miller CA, SH Settle, EP Sulman, KD Aldape, A Milosavljevic. Discovering
functional modules relevant for cancer progression by identifying patterns
of recurrent and mutually exclusive aberrations in tumor samples.
\emph{BMC Medical Genomics} 2011, \strong{4}:34.
}
\seealso{
\url{http://brl.bcm.tmc.edu/rme/index.rhtml}

\url{https://bmcmedgenomics.biomedcentral.com/articles/10.1186/1755-8794-4-34}
}
